requirements.txt
flask==2.3.3
pillow==10.0.0
flask-sqlalchemy==3.0.5
pymysql==1.1.0
cryptography==41.0.3
numpy==1.24.3
opencv-python==4.7.0.72
pywavelets==1.4.1
scipy==1.10.1